Removes more compiler warnings. And fixes actual errors with importing non-existent classes which Xcode told me I needed to import...

+++ ThunderCloud/TSCPokemonListItemView.m
@@ -176,7 +176,7 @@
 
 #pragma mark - TSCPokemonTableViewCellDelegate methods
 
-- (void)tableViewCell:(TSCPokemonTableViewCell *)cell didTapItemAtIndex:(int)index
+- (void)tableViewCell:(TSCPokemonTableViewCell *)cell didTapItemAtIndex:(NSInteger)index
 {
     TSCPokemonListItem *item = [self.items objectAtIndex:index];
